requirements.txt
langchain == 0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
# langchain_openai == 0.1.11  deprecated ()
langchain-openai==0.1.14
requests==2.32.3
openai==1.35.7
tiktoken==0.7.0
deepgram-brain==0.1.1
deepgram-sdk==3.3.1
# crewai-tools == 0.3.0
audio-recorder-streamlit==0.0.9
sounddevice==0.4.7
matplotlib==3.9.0
# dotenv==0.0.5
python-dotenv==1.0.1 #
scipy==1.14.0
